amended about us page
kinda changed it completely apart from the content.
made it look like it is part of the actual website as discussed on the wireframe during week 3.
added bootstrap link and also added bootstrap files in the CSS folder.
Still need to add the rest of the images at the bottom of the page.