homepage-navbar, image hover,contact-form amend
added navbar static on top of the page.
added hover on the images in teal to match the rest of the website.
added "video" tags on videos but still need to sort out the boxes as they overflow? and the buttons are huge.

Contact us page I amended the height of the form and amended the navbar for cohesion.
About us page i amended the navbar.